Split module graph JSON tests
